made a separate reusable widget for bottom shadow and refactored
